Fix 64 bit compile error.
git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk/nimbus@4232 640d5228-2204-0410-8d71-8f46fa6850b4

+++ src/filter/hdr_decode.c
@@ -2185,7 +2185,7 @@ static DERTBL *add_name_to_DERTBL(const char name[])
   if (strncmp(name, "FLX", 3) == 0)
     {
     dp->Output		= false;
-    dp->ProbeCount	= (int)deriveftns[indx].xlate;
+    dp->ProbeCount	= (size_t)deriveftns[indx].xlate;
     dp->ProbeType	= PROBE_FLUX;
     }
